Fix list of tuples. Closes #827 (#828)
