+++ tests/unit/lib/test_python_to_sql.py
@@ -1,0 +1,45 @@
+from collections.abc import Mapping
+from typing import Literal, Optional, Union
+
+import pytest
+
+from datachain.lib.convert.python_to_sql import python_to_sql
+from datachain.sql.types import JSON, Array, Float, String
+from tests.unit.lib.test_utils import MyModel
+
+
+@pytest.mark.parametrize(
+    "typ,expected",
+    (
+        (str, String),
+        (String, String),
+        (Literal["text"], String),
+        (dict[str, int], JSON),
+        (Mapping[str, int], JSON),
+        (Optional[str], String),
+        (Union[dict, list[dict]], JSON),
+    ),
+)
+def test_convert_type_to_datachain(typ, expected):
+    assert python_to_sql(typ) == expected
+
+
+def test_list_of_tuples_matching_types():
+    assert (
+        python_to_sql(list[tuple[float, float]]).to_dict()
+        == Array(Array(Float)).to_dict()
+    )
+
+
+def test_list_of_tuples_not_matching_types():
+    assert (
+        python_to_sql(list[tuple[float, String]]).to_dict()
+        == Array(Array(JSON)).to_dict()
+    )
+
+
+def test_list_of_tuples_object():
+    assert (
+        python_to_sql(list[tuple[float, MyModel]]).to_dict()
+        == Array(Array(JSON)).to_dict()
+    )
